requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.1
django-allauth==0.35.0
djangorestframework==3.7.7
idna==2.6
oauthlib==2.0.6
plaid-python==2.2.0
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22